Add support for OAUTH_URL and API_URL environment vars.
Set OAUTH_URL environment to override the URL used for OAuth token generation & exchanges.

Set API_URL to override the URL used for all API calls.

Usually these two are set together (though not to the same value) to refer to a development or testing stack as opposed to the main Onshape production stack.